updated test, compiling and running locally
